+++ module/item/item.mjs
@@ -688,7 +688,7 @@ export class HeroSystem6eItem extends Item {
 
         // If we have control of this token, re-acquire to update movement types
         const myToken = this.actor?.getActiveTokens()?.[0];
-        if (canvas.tokens.controlled.find((t) => t.id == myToken.id)) {
+        if (canvas.tokens.controlled?.find((t) => t.id == myToken.id)) {
             myToken.release();
             myToken.control();
         }
@@ -5289,6 +5289,19 @@ export class HeroSystem6eItem extends Item {
         }
 
         return RoundFavorPlayerDown(cost) + costSuffix;
+    }
+
+    /// Get Levels with AID/DRAIN Active Effects
+    get adjustedLevels() {
+        let _adjustedLevels = parseInt(this.system.LEVELS || 0);
+
+        for (const ae of this.actor.temporaryEffects.filter(
+            (effect) => effect.flags.XMLID === "DRAIN" && effect.flags.key === "POWERDEFENSE",
+        )) {
+            console.log(ae);
+            _adjustedLevels += parseInt(ae.changes?.[0].value || 0);
+        }
+        return Math.max(0, _adjustedLevels);
     }
 }