+++ module/utility/adjustment.mjs
@@ -216,7 +216,7 @@ function _findExistingMatchingEffect(item, potentialCharacteristic, powerTargetN
     return targetSystem.effects.find(
         (effect) =>
             effect.origin === item.uuid &&
-            effect.flags.target[0] === (potentialCharacteristic || powerTargetName?.uuid),
+            effect.flags.target[0] === (powerTargetName?.uuid || potentialCharacteristic),
     );
 }
 
@@ -263,13 +263,23 @@ function _createNewAdjustmentEffect(
     rawActivePointsDamage,
     targetActor,
     targetSystem,
+    action,
 ) {
     // Create new ActiveEffect
     // TODO: Add a document field
+
+    // Educated guess for token
+    const itemTokenName =
+        canvas.tokens.get(action?.current?.attackerTokenId)?.name ||
+        item.actor?.getActiveTokens().find((t) => canvas.tokens.controlled.find((c) => c.id === t.id))?.name ||
+        item.actor?.getActiveTokens()?.[0].name ||
+        targetActor.name ||
+        "undefined";
+
     const activeEffect = {
         name: `${item.system.XMLID || "undefined"} 0 ${
             (targetPower?.name || potentialCharacteristic)?.toUpperCase() // TODO: This will need to change for multiple effects
-        } (0 AP) [by ${item.actor.name || "undefined"}]`,
+        } (0 AP) [by ${itemTokenName}]`,
         id: `${item.system.XMLID}.${item.id}.${
             targetPower?.name || potentialCharacteristic // TODO: This will need to change for multiple effects
         }`,
@@ -285,8 +295,10 @@ function _createNewAdjustmentEffect(
             affectedPoints: 0,
             XMLID: item.system.XMLID,
             source: targetActor.name,
-            target: [potentialCharacteristic || targetPower?.uuid],
-            key: potentialCharacteristic,
+            target: [targetPower?.uuid || potentialCharacteristic],
+            key: targetPower?.system?.XMLID || potentialCharacteristic,
+            itemTokenName,
+            attackerTokenId: action?.current?.attackerTokenId,
         },
         origin: item.uuid,
         //description: item.system.description,  // Issues with core FoundryVTT where description doesn't show, nor is editable.
@@ -324,6 +336,7 @@ export async function performAdjustment(
     effectsDescription,
     isFade,
     targetActor,
+    action,
 ) {
     const isHealing = item.system.XMLID === "HEALING";
     const isOnlyToStartingValues = item.findModsByXmlid("ONLYTOSTARTING") || isHealing;
@@ -381,7 +394,9 @@ export async function performAdjustment(
 
     // Do we have a target?
     if (!targetCharacteristic && !targetPower) {
-        await ui.notifications.error(`${nameOfCharOrPower} is an invalid target for the adjustment power ${item.name}`);
+        await ui.notifications.warn(
+            `${nameOfCharOrPower} is an invalid target for the adjustment power ${item.name}. Perhaps ${targetActor.name} does not have that characteristic or power.`,
+        );
         return;
     }
 
@@ -408,6 +423,7 @@ export async function performAdjustment(
             thisAttackRawActivePointsDamage,
             targetActor,
             targetSystem,
+            action,
         );
 
     // Healing doesn't fade
@@ -548,10 +564,18 @@ export async function performAdjustment(
             parseInt(activeEffect.changes[2].value) + (newCalculatedValue - oldCalculatedValue);
     }
 
+    // Educated guess for token
+    const itemTokenName =
+        canvas.tokens.get(action?.current?.attackerTokenId)?.name ||
+        item.actor?.getActiveTokens().find((t) => canvas.tokens.controlled.find((c) => c.id === t.id))?.name ||
+        item.actor?.getActiveTokens()?.[0].name ||
+        targetActor.name ||
+        "undefined";
+
     // Update the effect max value(s)
     activeEffect.name = `${item.system.XMLID || "undefined"} ${Math.abs(totalActivePointsThatShouldBeAffected)} ${(
         targetPower?.name || potentialCharacteristic
-    )?.toUpperCase()} (${Math.abs(totalAdjustmentNewActivePoints)} AP) [by ${item.actor.name || "undefined"}]`;
+    )?.toUpperCase()} (${Math.abs(totalAdjustmentNewActivePoints)} AP) [by ${itemTokenName}]`;
 
     activeEffect.flags.affectedPoints = totalActivePointsThatShouldBeAffected;
     activeEffect.flags.adjustmentActivePoints = totalAdjustmentNewActivePoints;